Major refactoring of cfp admin table
