+++ .github/copilot-instructions.md
@@ -38,3 +38,47 @@ The site is multi-tenant, meaning it can be used for multiple events or conferen
 - Document any significant changes or features in the project README or relevant files under the /docs directory.
 - Prioritize performance and accessibility.
 - Use Vercel for deployment previews and production hosting.
+
+## Code Organization & Refactoring
+
+### Component Structure
+- Use a modular approach to components:
+  - Keep components focused on a single responsibility
+  - Break large components into smaller, reusable ones
+  - Place related components in dedicated subdirectories (e.g., `/components/proposal/`)
+  - Name components clearly to reflect their purpose (e.g., `ProposalTableRow`, `ProposalTableHeader`)
+
+### Business Logic
+- Extract complex business logic into custom hooks:
+  - Keep UI components focused on rendering
+  - Create hooks that handle state management and data manipulation (e.g., `useProposalSort`, `useProposalFilter`)
+  - Place hooks in a dedicated `/hooks` directory
+  - Name hooks with the `use` prefix followed by a descriptive name
+
+### Refactoring Approach
+When refactoring code:
+1. Identify overly complex components or duplicate functionality
+2. Determine logical separations of concerns
+3. Create new files/components for each distinct concern
+4. Extract shared logic into custom hooks or utility functions
+5. Ensure new components/hooks have clear interfaces with TypeScript types
+6. Update the original component to use the new structure
+
+### Folder Structure
+- Group related functionality in dedicated directories:
+  - `/components/{feature}/` - UI components for a specific feature
+  - `/hooks/` - Custom React hooks
+  - `/lib/{feature}/` - Feature-specific utility functions and types:
+    - `types.ts` - TypeScript type definitions specific to this feature
+    - `sanity.ts` - Sanity CMS queries and data fetching logic
+    - `client.ts` - Browser client-side code and API calls
+    - `server.ts` - Server-specific functions
+  - `/types/` - Shared TypeScript type definitions
+
+### Component Props
+- Define clear TypeScript interfaces for component props
+- Use consistent naming patterns (e.g., `{ComponentName}Props`)
+- Provide sensible defaults where appropriate
+- Document any complex or non-obvious props
+
+This structure ensures the codebase remains maintainable, testable, and scalable as the project grows.
